Added read_fasta() function
