+++ mokapot/__init__.py
@@ -22,7 +22,8 @@ except ImportError:
 from .dataset import LinearPsmDataset
 from .model import Model, PercolatorModel, save_model, load_model
 from .brew import brew
-from .parsers import read_pin, read_percolator, read_pepxml
+from .parsers.pin import read_pin, read_percolator
+from .parsers.pepxml import read_pepxml
+from .parsers.fasta import read_fasta, make_decoys, digest
 from .writers import to_flashlfq, to_txt
 from .confidence import LinearConfidence, plot_qvalues
-from .proteins import make_decoys, digest, FastaProteins
