+++ mokapot/parsers/__init__.py
@@ -1,5 +1,6 @@
 """
 Export parsers for various formats
 """
-from .pin import read_pin, read_percolator
-from .pepxml import read_pepxml
+# from .pin import read_pin, read_percolator
+# from .pepxml import read_pepxml
+# from .fasta import read_fasta
