+++ tests/system_tests/test_system.py
@@ -22,7 +22,7 @@ def test_compare_to_percolator():
 
     perc_path = os.path.join("data", "percolator.{l}.txt")
     perc_res = {
-        l: mokapot.parsers.read_percolator(perc_path.format(l=l))
+        l: mokapot.read_percolator(perc_path.format(l=l))
         for l in ["psms", "peptides", "proteins"]
     }
